api/requirements.txt
# CollectorStream API Dependencies
fastapi>=0.109.0
python-dotenv>=1.0.0
uvicorn[standard]>=0.27.0
pydantic[email]>=2.5.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.0
httpx>=0.26.0
boto3>=1.34.0
Pillow>=10.2.0
PyJWT>=2.8.0
